Moves the checks and tests to the main application and removes the old applications
