+++ pyproject.toml
@@ -269,8 +269,6 @@ testpaths = [
   "src/documents/tests/",
   "src/paperless/tests/",
   "src/paperless_mail/tests/",
-  "src/paperless_tesseract/tests/",
-  "src/paperless_remote/tests/",
   "src/paperless_ai/tests",
 ]
