--- src/paperless_remote/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PaperlessRemoteParserConfig(AppConfig):
-    name = "paperless_remote"
