--- src/paperless_tesseract/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PaperlessTesseractConfig(AppConfig):
-    name = "paperless_tesseract"
